requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.2.3
psycopg2==2.9.10
scikit-learn==1.6.1
scipy==1.15.2
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2025.1
pandas
scikit-learn
numpy
nltk
